Adicionado handlers para outras exceções

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,34 +3,40 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.models.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
 
     @Autowired
     private ProductRepository repository;
-    @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
-        return new ProductDTO(product);
-    }
 
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable){
-        Page<Product> products = repository.findAll(pageable);
-        return products.map(ProductDTO::new);
+        Page<Product> result = repository.findAll(pageable);
+        return result.map(ProductDTO::new);
     }
 
-    @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    @Transactional(readOnly = true)
+    public ProductDTO findById(Long id){
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Elemento não encontrado ou inexistente!"));
+        ProductDTO dto = new ProductDTO(product);
+        return dto;
+    }
+
+    @Transactional(readOnly = false)
+    public ProductDTO insert(ProductDTO dto) {
         Product product = new Product();
         product.setName(dto.getName());
         product.setDescription(dto.getDescription());
@@ -38,25 +44,36 @@ public class ProductService {
         product.setImgUrl(dto.getImgUrl());
 
         product = repository.save(product);
-
         return new ProductDTO(product);
     }
 
-    @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
-        Product product = repository.getReferenceById(id);
-        product.setName(dto.getName());
-        product.setDescription(dto.getDescription());
-        product.setPrice(dto.getPrice());
-        product.setImgUrl(dto.getImgUrl());
+    @Transactional(readOnly = false)
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
 
-        product = repository.save(product);
+            Product product = repository.getReferenceById(id);
+            product.setName(dto.getName());
+            product.setDescription(dto.getDescription());
+            product.setPrice(dto.getPrice());
+            product.setImgUrl(dto.getImgUrl());
 
-        return new ProductDTO(product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Elemento não encontrado ou inexistente!");
+        }
+
     }
-    @Transactional
+
+    @Transactional(readOnly = false, propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Elemento não encontrado ou inexistente!");
+        } try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
-
